+++ packages/ack/lib/src/schemas/schema.dart
@@ -24,6 +24,7 @@ part 'fluent_schema.dart';
 part 'instance_schema.dart';
 part 'lazy_schema.dart';
 part 'list_schema.dart';
+part 'map_schema.dart';
 part 'num_schema.dart';
 part 'object_schema.dart';
 part 'schema_type.dart';